--- feature/cover/src/main/java/com/my/version/feature/cover/CoverSecondViewModel.kt
@@ -1,10 +1,0 @@
-package com.my.version.feature.cover
-
-import androidx.lifecycle.ViewModel
-import javax.inject.Inject
-
-class CoverSecondViewModel @Inject constructor(
-
-): ViewModel() {
-
-}
