+++ feature/cover/src/main/java/com/my/version/feature/cover/main/navigation/CoverNavigation.kt
@@ -1,4 +1,4 @@
-package com.my.version.feature.cover.navigation
+package com.my.version.feature.cover.main.navigation
 
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
@@ -6,7 +6,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.my.version.core.common.navigation.MainTabRoute
-import com.my.version.feature.cover.CoverRoute
+import com.my.version.feature.cover.main.CoverRoute
 import kotlinx.serialization.Serializable
 
 fun NavController.navigateToCover(navOptions: NavOptions? = null) = navigate(Cover, navOptions)
