+++ feature/cover/src/main/java/com/my/version/feature/cover/select/navigation/CoverFirstNavigation.kt
@@ -1,4 +1,4 @@
-package com.my.version.feature.cover.navigation
+package com.my.version.feature.cover.select.navigation
 
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
@@ -6,7 +6,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.my.version.core.common.navigation.Route
-import com.my.version.feature.cover.CoverFirstRoute
+import com.my.version.feature.cover.select.CoverFirstRoute
 import kotlinx.serialization.Serializable
 
 fun NavController.navigateToCoverFirst(navOptions: NavOptions? = null) = navigate(CoverFirst, navOptions)
