+++ feature/cover/src/main/java/com/my/version/feature/cover/upload/CoverUploadViewModel.kt
@@ -1,0 +1,10 @@
+package com.my.version.feature.cover.upload
+
+import androidx.lifecycle.ViewModel
+import javax.inject.Inject
+
+class CoverUploadViewModel @Inject constructor(
+
+): ViewModel() {
+
+}
