+++ feature/cover/src/main/java/com/my/version/feature/cover/upload/navigation/CoverUploadNavigation.kt
@@ -1,4 +1,4 @@
-package com.my.version.feature.cover.navigation
+package com.my.version.feature.cover.upload.navigation
 
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
@@ -6,21 +6,20 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.my.version.core.common.navigation.Route
-import com.my.version.feature.cover.CoverFirstRoute
-import com.my.version.feature.cover.CoverSecondRoute
+import com.my.version.feature.cover.upload.CoverUploadRoute
 import kotlinx.serialization.Serializable
 
-fun NavController.navigateToCoverSecond(navOptions: NavOptions? = null) = navigate(CoverSecond, navOptions)
+fun NavController.navigateToCoverUpload(navOptions: NavOptions? = null) = navigate(CoverUpload, navOptions)
 
-fun NavGraphBuilder.coverSecondScreen(
+fun NavGraphBuilder.coverUploadScreen(
     modifier: Modifier
 ) {
-    composable<CoverSecond>{
-        CoverSecondRoute(
+    composable<CoverUpload>{
+        CoverUploadRoute(
             modifier = modifier
         )
     }
 }
 
 @Serializable
-data object CoverSecond: Route
+data object CoverUpload: Route
