+++ feature/main/src/main/java/com/my/version/feature/main/MainScreen.kt
@@ -3,13 +3,7 @@ package com.my.version.feature.main
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.foundation.layout.systemBarsPadding
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
@@ -24,17 +18,14 @@ import androidx.compose.ui.res.stringResource
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import com.my.version.core.common.navigation.Route
-import com.my.version.core.designsystem.component.topappbar.LogoTopAppBar
-import com.my.version.core.designsystem.theme.Grey200
-import com.my.version.core.designsystem.theme.Grey300
 import com.my.version.core.designsystem.theme.Grey400
 import com.my.version.core.designsystem.theme.MyVersionMain
 import com.my.version.core.designsystem.theme.White
 import com.my.version.feature.auth.signin.navigation.signInScreen
 import com.my.version.feature.auth.signup.navigation.signUpScreen
-import com.my.version.feature.cover.navigation.coverFirstScreen
-import com.my.version.feature.cover.navigation.coverScreen
-import com.my.version.feature.cover.navigation.coverSecondScreen
+import com.my.version.feature.cover.select.navigation.coverFirstScreen
+import com.my.version.feature.cover.main.navigation.coverScreen
+import com.my.version.feature.cover.upload.navigation.coverSecondScreen
 import com.my.version.feature.evaluate.main.navigation.evaluationScreen
 import com.my.version.feature.evaluate.record.navigation.evaluationRecordScreen
 import com.my.version.feature.evaluate.select.navigation.evaluationSelectScreen
