+++ feature/main/src/main/java/com/my/version/feature/main/MainTab.kt
@@ -5,7 +5,7 @@ import androidx.annotation.StringRes
 import androidx.compose.runtime.Composable
 import com.my.version.core.common.navigation.MainTabRoute
 import com.my.version.core.common.navigation.Route
-import com.my.version.feature.cover.navigation.Cover
+import com.my.version.feature.cover.main.navigation.Cover
 import com.my.version.feature.evaluate.main.navigation.Evaluation
 import com.my.version.feature.home.navigation.Home
